Added RegistraceForm, methods in RegistraceController and two templates: formular and shrnuti
RegistraceForm is Java Bean with attributes with validations. Two methods were added in RegistraceController: index and form. Template formular contains form to be filled in. Template shrnuti contains summary of data which were filled in the form.

Two things are not working. Firstly, there is no error message showing when the age is not between 9-14. Secondly, the summary is showing the code of a turnus but not the particular chosen dates.